--- frontend/src/pages/HomePage/components/ToolsSection/FeatureSections/index.ts
@@ -1,3 +1,0 @@
-export { default as FeatureCards } from './FeatureCards';
-export { default as FeatureCardsSection } from './FeatureCardsSection';
-export { default as FeatureDetailSections } from './FeatureCardsSections';
